refactor: extract top-level review package for evidence and run store
Move BookEvidenceIndex and ReviewRunStore out of pipeline into
trans_novel.review so agents depend on pure review models instead of
pipeline orchestration. Keep thin pipeline re-exports for compatibility
and lock the boundary with architecture tests.
