+++ tests/test_architecture_boundaries.py
@@ -5,7 +5,8 @@
   * 不使用 ThreadPoolExecutor / concurrent.futures；
   * 不直接调用 load_document / complete_json / build_report / assemble；
   * 装配全部六个拆分出的服务模块；
-  * 任何下层模块都不得反向导入 orchestrator.py。
+  * 任何下层模块都不得反向导入 orchestrator.py；
+  * agents 不得反向依赖 pipeline 编排模块（只允许依赖顶层 review 纯模型）。
 """
 
 from __future__ import annotations
@@ -14,7 +15,9 @@ import ast
 import pathlib
 import unittest
 
-PIPELINE_DIR = pathlib.Path(__file__).resolve().parent.parent / "trans_novel" / "pipeline"
+TRANS_NOVEL_DIR = pathlib.Path(__file__).resolve().parent.parent / "trans_novel"
+PIPELINE_DIR = TRANS_NOVEL_DIR / "pipeline"
+AGENTS_DIR = TRANS_NOVEL_DIR / "agents"
 
 SERVICE_MODULES = (
     "runtime",
@@ -37,9 +40,36 @@ FORBIDDEN_TOP_LEVEL = (
     "llm",
 )
 
+# agents 不得从 pipeline 拉取编排/状态机；review 纯模型在顶层 review/。
+FORBIDDEN_PIPELINE_MODULES_FOR_AGENTS = (
+    "orchestrator",
+    "runtime",
+    "preparation",
+    "annotations",
+    "translation",
+    "review_workflow",
+    "finalization",
+    "runstore",
+    "metrics",
+    "context",
+    "language",
+    "checks",
+    # 兼容 re-export 也禁止；agents 应直接依赖顶层 review/
+    "review_evidence",
+    "review_run",
+)
+
 
 def _module_source(name: str) -> str:
     return (PIPELINE_DIR / f"{name}.py").read_text(encoding="utf-8")
+
+
+def _agent_sources() -> list[tuple[str, str]]:
+    return [
+        (path.name, path.read_text(encoding="utf-8"))
+        for path in sorted(AGENTS_DIR.glob("*.py"))
+        if path.name != "__init__.py"
+    ]
 
 
 class TestArchitectureBoundaries(unittest.TestCase):
@@ -146,6 +176,47 @@ class TestArchitectureBoundaries(unittest.TestCase):
             self.assertTrue(hasattr(module, class_name), f"{module_name}.{class_name} 缺失")
         finalization = importlib.import_module("trans_novel.pipeline.finalization")
         self.assertTrue(hasattr(finalization, "AssemblyService"))
+
+    def test_agents_do_not_import_pipeline_orchestration(self):
+        """agents 只可依赖顶层 review 纯模型，不得倒挂 pipeline 编排模块。"""
+        for filename, source in _agent_sources():
+            tree = ast.parse(source)
+            for node in ast.walk(tree):
+                if not isinstance(node, ast.ImportFrom):
+                    continue
+                module = node.module or ""
+                parts = module.split(".")
+                if node.level == 2 and parts and parts[0] == "pipeline":
+                    rest = parts[1:] if len(parts) > 1 else []
+                    if not rest:
+                        self.fail(f"{filename} 不得 import ..pipeline")
+                    self.assertNotIn(
+                        rest[0],
+                        FORBIDDEN_PIPELINE_MODULES_FOR_AGENTS,
+                        f"{filename} 不得反向依赖 pipeline.{rest[0]}",
+                    )
+                if node.level == 0 and module.startswith("trans_novel.pipeline"):
+                    parts = module.split(".")
+                    if len(parts) >= 3:
+                        self.assertNotIn(
+                            parts[2],
+                            FORBIDDEN_PIPELINE_MODULES_FOR_AGENTS,
+                            f"{filename} 不得反向依赖 {module}",
+                        )
+
+    def test_review_package_exports_core_types(self):
+        """顶层 review 包提供证据索引与运行目录模型。"""
+        import importlib
+
+        review = importlib.import_module("trans_novel.review")
+        for name in (
+            "BookEvidenceIndex",
+            "SegmentRef",
+            "ReviewRunStore",
+            "ReviewOutcome",
+            "review_candidate_id",
+        ):
+            self.assertTrue(hasattr(review, name), f"trans_novel.review.{name} 缺失")
 
 
 if __name__ == "__main__":
